week5/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
httpx>=0.24.0
